appveyor: Add a wrap for llvm
The appveyor build has a copy of llvm 5.0.1 that scons uses, meson can
also use this to build llvm pipe (though not swr, which requires 6.0.0)
as well.

This requires a plain buildtype, since we need to compile single
threaded to link with the provided LLVM, as well as a 32 bit build for
the same reason.

v4: - New in this version
v5: - use ne b_vscrt option instead of manually specifying crt